orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "adaptive-os"
version = "0.1.0"
description = "AI-driven OS orchestrator that adapts the system to user context"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "psutil>=5.9",
    "aiohttp>=3.9",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
    "httpx>=0.27",
    "watchdog>=4.0",
    "pydantic>=2.0",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
adaptive-os = "adaptive_os.cli:cli"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
